+++ packages/mask/dashboard/pages/SetupPersona/Recovery/Cloud/PhoneField.tsx
@@ -10,11 +10,11 @@ import { useOutletContext } from 'react-router-dom'
 import { useAsyncFn } from 'react-use'
 import { useLanguage } from '../../../../../shared-ui/index.js'
 import { PrimaryButton } from '../../../../components/PrimaryButton/index.js'
-import { RestoreContext } from '../../../../components/Restore/RestoreFromCloud/RestoreProvider.js'
 import { sendCode, type RestoreQueryError } from '../../../../utils/api.js'
 import { phoneRegexp } from '../../../../utils/regexp.js'
 import { Locale, Scenario } from '../../../../utils/type.js'
 import type { PortalContainerProps } from '../../types.js'
+import { RestoreContext } from './RestoreProvider.js'
 
 export const PhoneField = memo(function PhoneField() {
     const { t } = useLingui()
@@ -100,6 +100,7 @@ export const PhoneField = memo(function PhoneField() {
                 <PrimaryButton
                     color="primary"
                     size="large"
+                    variant="roundedContained"
                     onClick={async () => {
                         dispatch({ type: 'SET_LOADING', loading: true })
                         try {
