+++ packages/mask/dashboard/pages/SetupPersona/Recovery/Cloud/RestoreProvider.tsx
@@ -1,8 +1,8 @@
 import { useCallback, useReducer } from 'react'
 import { createContainer } from '@masknet/shared-base-ui'
-import { fetchDownloadLink } from '../../../utils/api.js'
 import type { BackupAccountType } from '@masknet/shared-base'
 import { initialState, restoreReducer } from './restoreReducer.js'
+import { fetchDownloadLink } from '../../../../utils/api.js'
 
 function useRestoreState() {
     const [state, dispatch] = useReducer(restoreReducer, initialState)
@@ -13,6 +13,5 @@ function useRestoreState() {
     return { state, dispatch, downloadBackupInfo }
 }
 
-/** @deprecated */
 export const RestoreContext = createContainer(useRestoreState)
 RestoreContext.Provider.displayName = 'RestoreContextProvider'
