+++ packages/mask/dashboard/pages/SetupPersona/Recovery/Cloud/restoreReducer.ts
@@ -1,9 +1,8 @@
 import type { BackupSummary } from '@masknet/backup-format'
 import { produce } from 'immer'
-import { type BackupFileInfo } from '../../../utils/type.js'
 import { BackupAccountType } from '@masknet/shared-base'
+import type { BackupFileInfo } from '../../../../utils/type.js'
 
-/** @deprecated */
 export enum RestoreStep {
     InputEmail = 'InputEmail',
     InputPhone = 'InputPhone',
