+++ packages/mask/dashboard/pages/SetupPersona/Recovery/Local.tsx
@@ -200,6 +200,7 @@ export const Component = memo(function RecoveryLocalBackup() {
                 <PrimaryButton
                     size="large"
                     color="primary"
+                    variant="roundedContained"
                     onClick={restoreStatus === RestoreStatus.Decrypting ? decryptBackupFile : restoreDB}
                     loading={loading}
                     disabled={disabled}>
