+++ packages/mask/dashboard/pages/SetupPersona/Recovery/PrivateKey.tsx
@@ -101,6 +101,7 @@ export const Component = memo(function RecoveryPrivateKey() {
                 <PrimaryButton
                     size="large"
                     color="primary"
+                    variant="roundedContained"
                     disabled={isSubmitting || !isDirty}
                     onClick={handleSubmit(onSubmit)}>
                     <Trans>Continue</Trans>
